requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

ipython==8.12.3
keras==3.9.2
matplotlib==3.9.2
moviepy==2.1.2
numpy==1.26.4
opencv-python==4.11.0.86
pillow==10.4.0
streamlit==1.45.0
tensorflow==2.19.0
ultralytics==8.3.111
